order_service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.20.0
confluent-kafka>=2.0.0
python-dotenv>=1.0.0
